+++ packages/@expo/cli/src/export/embed/exportEmbedAsync.ts
@@ -125,6 +125,7 @@ export async function exportEmbedBundleAndAssetsAsync(
         // Never output bytecode in the exported bundle since that is hardcoded in the native run script.
         bytecode: false,
         // source map inline
+        reactCompiler: !!exp.experiments?.reactCompiler,
       },
       {
         sourceMapUrl,