+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -38,6 +38,7 @@ type Options = {
   entryPoint?: string;
   clear: boolean;
   routerRoot: string;
+  reactCompiler: boolean;
   maxWorkers?: number;
   isExporting: boolean;
 };